% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTerms.R
\name{getTerms}
\alias{getTerms}
\title{Compares a date to the terms of the Dartmouth academic calendar from 2009 - 2017}
\usage{
getTerms(date_time)
}
\arguments{
\item{date_time}{A valid date or datetime, i.e., 2016-08-01}
}
\description{
Takes a date or datetime and compares it do the dates of the Dartmouth academic calendar, returns a Term name.
}
\examples{
getTerms("2016-05-17")
df$terms <- getTerms(df$time)
}

