% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{suma_decode_activities}
\alias{suma_decode_activities}
\title{Decode activities from numbers into activity names}
\usage{
suma_decode_activities(df, key)
}
\arguments{
\item{df}{Data frame containing the initial Suma data}

\item{key}{Data frame containing the key table}
}
\description{
This function separates multiple activities attached to single Suma counts, compares them to a key table, and assigns activity names as appropriate.
}
\examples{
df <- suma_decode_activities(df, sumaKey)
}

