% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{suma_from_api}
\alias{suma_from_api}
\title{Gather data from the Suma API}
\usage{
suma_from_api(url = "https://library.dartmouth.edu/sumaserver/",
  initiativeId = 1, startDate = cut(Sys.Date(), "month"),
  endDate = Sys.Date(), sepDates = TRUE, unnest = TRUE)
}
\arguments{
\item{url}{The base url, defaults to Dartmouth's https://library.dartmouth.edu/suma/}

\item{initiativeId}{The desired initiative to retrieve data for, defaults to 1}

\item{startDate}{Desired start date, defaults to beginning of current month}

\item{endDate}{Desired end date, defaults to today's date}

\item{sepDates}{Creates columns for easy grouping and sorting of date and time from sessionStart}

\item{unnest}{If TRUE, separates counts with multiple activities into one row per activity}
}
\description{
This function hits the Suma API for the selected initiative ID during the given date range.
}
\examples{
df <- suma_from_api("https://library.dartmouth.edu/sumaserver/", 1, "2016-04-01", "2016-08-01")
}

