% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activities.R
\name{suma_get_activities}
\alias{suma_get_activities}
\title{Retrieves activity names from Suma server}
\usage{
suma_get_activities(url = "https://library.dartmouth.edu/sumaserver/")
}
\arguments{
\item{url}{The url of the Suma query server. If your client is at "yourlibrary.edu/suma/client" this is usually "yourlibrary.edu/sumaserver"}
}
\description{
This function retrieves a table of activity names and ID numbers, suitable for use as a key for suma_decode_activities function.
}
\examples{
sumaKey <- suma_get_activities()
}

