% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activities.R
\name{suma_get_initiatives}
\alias{suma_get_initiatives}
\title{Retrieves initiative names from Suma server}
\usage{
suma_get_initiatives(url = "https://library.dartmouth.edu/sumaserver/")
}
\arguments{
\item{url}{The url of the Suma query server. If your client is at "yourlibrary.edu/suma/client" this is usually "yourlibrary.edu/sumaserver"}
}
\description{
This function retrieves a table of initiative names and ID numbers, useful for knowing which intiative ID to use when calling suma_from_api.
}
\examples{
initiatives <- suma_get_initiatives()
}

