% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activities.R
\name{suma_get_locations}
\alias{suma_get_locations}
\title{Retrieves location names from Suma server}
\usage{
suma_get_locations(url = "https://library.dartmouth.edu/sumaserver/")
}
\arguments{
\item{url}{The url of the Suma query server. If your client is at "yourlibrary.edu/suma/client" this is usually "yourlibrary.edu/sumaserver"}
}
\description{
This function retrieves a table of location names and ID numbers.
}
\examples{
locs <- suma_get_locations()
}

