% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{suma_group_check}
\alias{suma_group_check}
\title{Helper function, checks whether to group results of Suma math functions and returns vector for use later}
\usage{
suma_group_check(groupLocation = FALSE, groupHour = FALSE,
  groupWeekday = FALSE, groupMonth = FALSE, groupActs = FALSE)
}
\arguments{
\item{groupLocation}{Optional argument to group data by location}

\item{groupHour}{Optional argument to group data by hour of day}

\item{groupWeekday}{Optional argument to group data by day of the week}

\item{groupMonth}{Optional argument to group data by month}

\item{groupActs}{Optional argument to group data by activity}
}
\description{
Helper function, checks whether to group results of Suma math functions and returns vector for use later
}

