% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{suma_math_conditions}
\alias{suma_math_conditions}
\title{Helper function to suma_MATH_count}
\usage{
suma_math_conditions(df, groupBy = NULL, filterBy = NULL, op)
}
\arguments{
\item{df}{Data frame containing Suma Data}

\item{groupBy}{Optional argument to group data, such as term, weekday, time}

\item{filterBy}{Optional argument to filter results}

\item{op}{Used to specify which operation will be used (min, max, mean, etc.)}
}
\description{
Helper function to suma_MATH_count
}

