% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{suma_math_filter}
\alias{suma_math_filter}
\title{Helper function to suma_MATH_count}
\usage{
suma_math_filter(df, filterBy, op)
}
\arguments{
\item{df}{Data frame containing Suma Data}

\item{filterBy}{Optional argument to filter results}
}
\description{
Helper function to suma_MATH_count
}

