% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{suma_math_group}
\alias{suma_math_group}
\title{Helper function to suma_MATH_count}
\usage{
suma_math_group(df, groupBy, op)
}
\arguments{
\item{df}{Data frame containing Suma Data}

\item{groupBy}{Optional argument to group data, such as term, weekday, time}
}
\description{
Helper function to suma_MATH_count
}

