% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{suma_math_nulls}
\alias{suma_math_nulls}
\title{Helper function to suma_MATH_count}
\usage{
suma_math_nulls(df, op)
}
\arguments{
\item{df}{Data frame containing Suma Data}
}
\description{
Helper function to suma_MATH_count
}

