% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{suma_max_count}
\alias{suma_max_count}
\title{Determine the highest actual count from a single session}
\usage{
suma_max_count(df, groupBy = NULL, filterBy = NULL)
}
\arguments{
\item{df}{Data frame containing Suma Data}

\item{groupBy}{Optional argument to group data, such as term, weekday, time}

\item{filterBy}{Optional argument to filter results}
}
\description{
This function determines the maximum count in a single session, with the option to group results
}

