% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{suma_max_group}
\alias{suma_max_group}
\title{Helper function to suma_max_count}
\usage{
suma_max_group(df, groupBy)
}
\arguments{
\item{df}{Data frame containing Suma Data}

\item{groupBy}{Optional argument to group data, such as term, weekday, time}
}
\description{
Helper function to suma_max_count
}

