% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{suma_max_nulls}
\alias{suma_max_nulls}
\title{Helper function to suma_max_count}
\usage{
suma_max_nulls(df)
}
\arguments{
\item{df}{Data frame containing Suma Data}
}
\description{
Helper function to suma_max_count
}

