% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{suma_mean_count}
\alias{suma_mean_count}
\title{Determine the mean count across sessions}
\usage{
suma_mean_count(df, filterBy = NULL, groupBy = NULL)
}
\arguments{
\item{df}{Data frame containing Suma Data}

\item{filterBy}{Optional argument to filter results}

\item{groupBy}{Optional argument to group data, such as term, weekday, time}
}
\description{
This function determines the mean count across sessions, with the option to group or filter results
}

