% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\docType{data}
\name{sumaKey}
\alias{sumaKey}
\title{Suma activity name data from Dartmouth}
\format{An object of class \code{data.frame} with 32 rows and 2 columns.}
\usage{
data("sumaKey")
}
\description{
Suma activity name data from Dartmouth
}
\keyword{datasets}

