% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{suma_group_check}
\alias{suma_group_check}
\title{Helper function, checks whether to group results of Suma math functions and returns vector for use later}
\usage{
suma_group_check(groupLocation = FALSE, groupHour = FALSE,
  groupWeekday = FALSE, groupMonth = FALSE, groupActs = FALSE)
}
\arguments{
\item{groupLocation}{Optional argument to group data by location}
}
\description{
Helper function, checks whether to group results of Suma math functions and returns vector for use later
}

