% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{suma_max_count}
\alias{suma_max_count}
\title{Determine the highest actual count across sessions}
\usage{
suma_max_count(df, filterBy = NULL, groupBy = NULL)
}
\arguments{
\item{df}{Data frame containing Suma Data}

\item{filterBy}{Optional argument to filter results}

\item{groupBy}{Optional argument to group data, such as term, wday, time}
}
\description{
This function determines the maximum count across sessions, with the option to group or filter results
}

