% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{suma_mean_count}
\alias{suma_mean_count}
\title{Determine the mean count from a single session}
\usage{
suma_mean_count(df, groupBy = NULL, filterBy = NULL)
}
\arguments{
\item{df}{Data frame containing Suma Data}

\item{groupBy}{Optional argument to group data, such as term, weekday, time}

\item{filterBy}{Optional argument to filter results}
}
\description{
This function determines the mean count across, with the option to group or filter results
}

