% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{suma_plot}
\alias{suma_plot}
\title{Quickly plot Suma data using ggplot2}
\usage{
suma_plot(df, xVar, op = "max", groupBy = NULL, filterBy = NULL,
  facetBy = NULL, fillBy = NULL)
}
\arguments{
\item{df}{Data frame containing Suma Data}

\item{xVar}{What variable you'd like to use for the X axis of your plot(s)}

\item{op}{What operation you want to use to get the data}

\item{groupBy}{Optional variable defining how to group data}

\item{filterBy}{Optional variable defining how to filter data}

\item{facetBy}{Optional variable for faceting your plot}

\item{fillBy}{Optional variable to define the fill of the bars}
}
\description{
Quickly plot Suma data with a pre-defined bar chart
}

